Fixed Client JSON parsing

+++ client/src/GameObject.java
@@ -2,9 +2,9 @@ package client;
 import java.util.*;
 
 public class GameObject {
-    public Hashtable<Integer,DrawObject> actors;
+    public Hashtable<Long,DrawObject> actors;
 
     public GameObject() {
-        actors = new Hashtable<Integer,DrawObject>();
+        actors = new Hashtable<Long,DrawObject>();
     }
 }
